fix: correct SQL errors in mart_aqi_daily and mart_neighborhood_comparison (closes #93) (#94)
- build_aqi_daily: wrap GROUP BY in derived table so correlated subquery
  references d.date instead of ungrouped s.observed_at
- build_neighborhood_comparison: add ::numeric cast to ROUND arguments
  for all three delta_pct calculations

+++ data/marts/build_neighborhood_comparison.py
@@ -34,13 +34,13 @@ SELECT
          THEN ROUND((COALESCE(cur.r311, 0)::numeric / n.shape_area) * 1000000, 4)
          ELSE 0 END,
     CASE WHEN COALESCE(prev.crime, 0) > 0
-         THEN ROUND(((COALESCE(cur.crime, 0) - prev.crime)::numeric / prev.crime) * 100, 1)
+         THEN ROUND(((COALESCE(cur.crime, 0) - prev.crime)::numeric / prev.crime * 100)::numeric, 1)
          ELSE NULL END,
     CASE WHEN COALESCE(prev.crashes, 0) > 0
-         THEN ROUND(((COALESCE(cur.crashes, 0) - prev.crashes)::numeric / prev.crashes) * 100, 1)
+         THEN ROUND(((COALESCE(cur.crashes, 0) - prev.crashes)::numeric / prev.crashes * 100)::numeric, 1)
          ELSE NULL END,
     CASE WHEN COALESCE(prev.r311, 0) > 0
-         THEN ROUND(((COALESCE(cur.r311, 0) - prev.r311)::numeric / prev.r311) * 100, 1)
+         THEN ROUND(((COALESCE(cur.r311, 0) - prev.r311)::numeric / prev.r311 * 100)::numeric, 1)
          ELSE NULL END,
     NOW()
 FROM stg_neighborhoods n
